miscellaneous/customer-relationship-management/requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
Brotli==1.1.0
bs4==0.0.2
celery==5.3.6
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==2023.10
cron-descriptor==1.4.3
cssselect2==0.7.0
Django==4.2.9
django-appconf==1.0.6
django-auditlog==2.3.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-countries==7.5.1
django-crispy-forms==2.1
django-datatables-view==1.20.0
django-extensions==3.2.3
django-js-asset==2.2.0
django-phone-field==1.8.1
django-phonenumber-field==7.3.0
django-select2==8.1.2
django-tables2==2.7.0
django-taggit==5.0.1
django-timezone-field==6.1.0
django-weasyprint==2.2.2
djangorestframework==3.14.0
fonttools==4.47.2
frozenlist==1.4.1
html5lib==1.1
idna==3.6
kombu==5.3.5
multidict==6.0.5
mysqlclient==2.2.3
numpy==1.26.4
ortools==9.8.3296
pandas==2.2.0
phonenumbers==8.13.30
phonenumberslite==8.13.30
pillow==10.2.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psycopg2-binary==2.9.9
pycparser==2.21
pydyf==0.8.0
pymemcache==4.0.0
pyphen==0.14.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2024.1
schedule==1.2.1
setuptools==69.0.3
six==1.16.0
soupsieve==2.5
split-settings==1.0.0
sqlparse==0.4.4
tinycss2==1.2.1
typing_extensions==4.9.0
tzdata==2023.4
vine==5.1.0
wcwidth==0.2.13
weasyprint==60.2
webencodings==0.5.1
yarl==1.9.4
zopfli==0.2.3
